pyproject.toml
[tool.poetry]
name = "DicePP"
version = "1.0.0 Beta2"
description = "DicePP"
authors = ["Pear"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
aiohttp = "^4.0.2"
nonebot2 = "^2.0.0b1"
nb-cli = "^0.6.4"
openpyxl = "^3.0.9"
rsa = "^4.8"
nonebot-adapter-cqhttp = "^2.0.0b1"
gitpython = "3.1.26"

[nonebot.plugins]
plugins = []
plugin_dirs = ["src/plugins"]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[[tool.poetry.source]]
name = "tsinghua"
default = true
url = "https://pypi.tuna.tsinghua.edu.cn/simple"